Auth callback: move client logic to CallbackRunner, page is server with Suspense + force-dynamic
